move metric controller from app->app harvest
